chore: move BKTValue & BKTEvaluationDetails to new files
