+++ src/BKTEvaluationDetails.ts
@@ -1,14 +1,12 @@
-/**
- * @deprecated use BKTEvaluationDetails<T> instead.
- */
-export interface BKTEvaluation {
-  readonly id: string
+import { BKTValue } from './BKTValue'
+
+export interface BKTEvaluationDetails<T extends BKTValue> {
   readonly featureId: string
   readonly featureVersion: number
   readonly userId: string
   readonly variationId: string
   readonly variationName: string
-  readonly variationValue: string
+  readonly variationValue: T
   readonly reason:
     | 'TARGET'
     | 'RULE'
