+++ src/main.ts
@@ -15,16 +15,17 @@ export { defineBKTUser } from './BKTUser'
 export type { BKTClient } from './BKTClient'
 export { getBKTClient, destroyBKTClient } from './BKTClient'
 export type {
-  BKTValue,
-  BKTJsonArray,
-  BKTJsonObject,
-  BKTJsonPrimitive,
-} from './JsonTypes'
-export type {
   BKTStorage,
   BrowserLocalStorage,
   InMemoryStorage,
 } from './BKTStorage'
+export type {
+  BKTValue,
+  BKTJsonArray,
+  BKTJsonObject,
+  BKTJsonPrimitive,
+} from './BKTValue'
+export type { BKTEvaluationDetails } from './BKTEvaluationDetails'
 
 const createNodeComponent = (config: BKTConfig, user: User): Component => {
   return new DefaultComponent(
